Check for table existence for vault (see note)
This was the last revision to the database that was handled with
SQLAlchemy Migrate prior to the move to Alembic (db version 180).
However, it was added to the dev branch after 21.09 was released, but
before 22.01. Therefore, this table should be created when upgrading
from 21.09, but not from dev.

+++ lib/galaxy/model/migrations/alembic/versions_gxy/40aaada107df_create_table_vault.py
@@ -9,6 +9,7 @@ import datetime
 
 import sqlalchemy as sa
 from alembic import op
+from sqlalchemy.engine.reflection import Inspector
 
 
 # revision identifiers, used by Alembic.
@@ -19,15 +20,26 @@ depends_on = None
 
 
 def upgrade():
-    now = datetime.datetime.utcnow
-    op.create_table('vault',
-        sa.Column('key', sa.Text, primary_key=True),
-        sa.Column('parent_key', sa.Text, sa.ForeignKey('vault.key'), index=True, nullable=True),
-        sa.Column('value', sa.Text, nullable=True),
-        sa.Column("create_time", sa.DateTime, default=now),
-        sa.Column("update_time", sa.DateTime, default=now, onupdate=now),
-    )
+    # This is the last revision to the database that was handled with SQLAlchemy Migrate prior to
+    # the move to Alembic. However, it was added to the dev branch after 21.09 was released, but
+    # before 22.01. Therefore, this table should be created when upgrading from 21.09, but not from dev.
+    if not _table_exists('vault'):
+        now = datetime.datetime.utcnow
+        op.create_table('vault',
+            sa.Column('key', sa.Text, primary_key=True),
+            sa.Column('parent_key', sa.Text, sa.ForeignKey('vault.key'), index=True, nullable=True),
+            sa.Column('value', sa.Text, nullable=True),
+            sa.Column("create_time", sa.DateTime, default=now),
+            sa.Column("update_time", sa.DateTime, default=now, onupdate=now),
+        )
 
 
 def downgrade():
-    op.drop_table('vault')
+    # See comment in upgrade function
+    if _table_exists('vault'):
+        op.drop_table('vault')
+
+
+def _table_exists(table: str):
+    inspector = Inspector.from_engine(op.get_bind())
+    return table in inspector.get_table_names()
